requirements.txt
# Core Flask dependencies
Flask==3.0.3
Werkzeug==3.0.4
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.8.2

# Database ORM and Management
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Password hashing and security
bcrypt==4.1.2
flask-bcrypt==1.0.1

# Configuration and Environment
python-dotenv==1.0.1

# Development and Testing dependencies
pytest==8.3.2
pytest-cov==5.0.0
pytest-flask==1.3.0

# Code Quality and Formatting
flake8==7.1.1
black==24.8.0
isort==5.13.2

# Security and Validation
marshmallow==3.21.3
flask-marshmallow==1.2.1

# CORS support (useful for frontend integration)
flask-cors==4.0.1

# Production WSGI server
gunicorn==22.0.0

# HTTP requests (for testing external APIs)
requests==2.32.3

# JSON Web Token support (future enhancement)
PyJWT==2.8.0

# Date and time utilities
python-dateutil==2.8.2

# UUID utilities
uuid==1.30

# API Documentation
flasgger==0.9.7.1
flask-restx==1.3.0
PyYAML==6.0.1
